backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
requests==2.26.0
psycopg2-binary==2.9.3
PyJWT==2.1.0
requests==2.26.0
pillow==10.0.1
